List works (needed to be in MainActivity) Bug: onClick() for list elements crashes.
