+++ app/src/main/java/fr3380ot/com/gro2/AddHabit.java
@@ -3,11 +3,13 @@ package fr3380ot.com.gro2;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 
 /**
@@ -62,6 +64,11 @@ public class AddHabit extends Activity {
         Intent intent = new Intent(getApplication(), MainActivity.class);
         startActivity(intent);
         this.finish();
+    }
+
+    public void test1(View view) {
+        ArrayList<HashMap<String, String>> test = dbTools.getAllHabits();
+        Log.d("Habit = ", test.toString());
     }
 
 }
