+++ app/src/main/java/fr3380ot/com/gro2/MainActivity.java
@@ -13,7 +13,10 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.widget.AdapterView;
+import android.widget.ListAdapter;
 import android.widget.ListView;
+import android.widget.SimpleAdapter;
 import android.widget.TextView;
 
 import java.util.ArrayList;
@@ -21,6 +24,10 @@ import java.util.HashMap;
 
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
+
+    DBTools dbTools = new DBTools(this);
+    Intent intent;
+    TextView habitId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -52,6 +59,42 @@ public class MainActivity extends AppCompatActivity
 
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
+
+
+
+
+
+
+        //Pull habits from database and store in habitList
+        ArrayList<HashMap<String, String>> habitList = dbTools.getAllHabits();
+        dbTools.getAllHabits();
+
+        //If there are habits created
+        if(habitList.size() != 0) {
+            ListView listView = (ListView) findViewById(android.R.id.list);
+            listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+
+                // When an item is clicked, get the TextView with the matching Id
+                @Override
+                public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                    //the item clicked
+                    habitId = (TextView) view.findViewById(R.id.habitId);
+                    String habitIdValue = habitId.getText().toString();
+
+                    intent = new Intent(getApplication(), EditHabit.class);
+
+                    intent.putExtra("habitId", habitIdValue);
+
+                    startActivity(intent);
+                }
+            });
+
+            ListAdapter adapter = new SimpleAdapter(
+                    this, habitList, R.layout.habit_entry, new String[]{"habitId", "title"}, new int[]{
+                    R.id.habitId, R.id.habitTitle});
+
+            listView.setAdapter(adapter);
+        }
     }
 
     @Override
