+++ app/src/main/res/layout/add_habit.xml
@@ -102,4 +102,16 @@
             android:layout_weight="1"/>
     </TableRow>
 
+    <TableRow
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="New Button"
+            android:onClick="test1"
+            android:id="@+id/button2" />
+    </TableRow>
+
 </TableLayout>
